tests/requirements.txt
# Test dependencies
pytest>=7.0.0
psycopg2-binary>=2.9.0  # Postgres driver (binary = no compile needed)
redis>=4.0.0            # Redis driver
